feat(tictactoe): initial implementation of game logic
This is a really simple implementation of a Tic-Tac-Toe game, which is being used as part of a demo
for my jQueryTO presentation. Not a whole lot of effort has been put into setting this up, and there's
no real documentation or anything, as well as nothing interesting like an AI or particularly useful
set of rules.

If anyone is ever interested in improving this, for whatever reason, they are of course more than
welcome to have a go at it.